Missing: msm: kgsl: Add CFF user event ioctl to write data to cff dumps
and

msm: kgsl: add MEMORY_BASE and HANG packets

These are changes to the cffdump output format.
MEMORY_BASE reports the address range we're using at the start
of the dump, which helps the postprocessors validate memory
writes.

HANG is emitted at the start of kgsl_postmortem_dump(), and can
be used by the postprocessors to stop when the hang occurs.

Change-Id: Ia60557d34bcaf3a20b3852f880a27f3605fecad2
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>