driver: usb: add CSW Hack feature
squashed commit of:

Implementing the csw hack feature & increasing the number of
buffers used in mass storage to 4.

As per the mass storage protocol, csw is sent from the device
after successfully writing the data on to storage media, where
there is a time factor involved. With this csw hack, csw is sent
just after receiving the data & before writing the data on the
storage media. By, any chance, if the write fails, the corresponding
lun is removed to indicate the user of write failure.

Buffers has been increased to 4, so that there is enough space for
the incoming data, which is usually 64KB per CBW.

User has the option of selecting/deselecting this feature from
the kernel config.

Signed-off-by: Velempati Chiranjeevi <c_cvelem@qualcomm.com>

----

USB: function: mass_storage: Fix in the CSW Hack feature

Before sending the csw to the host, need to cross check,
whether we have received the complete intended data from
the host.

CRs-Fixed: 182952
CRs-Fixed: 182411

Signed-off-by: Velempati, Chiranjeevi <c_cvelem@qualcomm.com>

----

USB: f_mass_storage: Fix multiple enumeration, if media removed abruptly

As part of CSW HACK feature, sending the CSW before actually writing the
data on to the media. If the media is removed after sending the CSW and
before the data is actually written, then the host will not know the
write failure. Hence send the write failure info in the next CSW, so
that host knows about the write failure and throw the right message to
the user.

CRs-fixed: 267925
Change-Id: Ie3d58fb33f26e2f563ecb4560b3fd1249377b091
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>

----

USB: f_mass_storage: Fix sending of CSW if the WRITE command fails

With CSW_HACK feature, CSW is returned immediately for WRITE commands
without waiting for data to be written to the media. There are some
scenarios where because of some error while processing of WRITE command
(e.g unmounting of media during data transfer), CSW is not sent.
Handle all such corner cases.

Change-Id: Id795b7ec7afc10e67a3309e6b650dfc5b7867c9e
CRs-fixed: 281801
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>

Conflicts:

	drivers/usb/gadget/Kconfig
	drivers/usb/gadget/f_mass_storage.c
	drivers/usb/gadget/storage_common.c